+++ lib/src/core/common/platform_manager.dart
@@ -23,6 +23,23 @@ class DefaultsPlatformManager {
 
   /// Initializes the DefaultsPlatformManager with the specified parameters.
   ///
+  /// - [targetPlatform] Use MobileTargetPlatform.<value> for the application:
+  ///
+  /// Example usage:
+  /// ```dart
+  /// void main() async {
+  ///   DefaultsPlatformManager.initialize(
+  ///     MobileTargetPlatform.android,
+  ///     isDebugging: true,
+  ///   );
+  ///   runApp(const App());
+  /// }
+  /// ```
+  ///
+  /// - [targetWeb] is the target web platform or defaults to [targetPlatform] if not specified.
+  /// - [isDebugging] is the debugging status for the application, If set to false,
+  /// the [targetPlatform] parameter will be ignored, and the specific widget behavior will depend on the base platform.
+  ///
   /// Throws an error if the manager is already initialized.
   factory DefaultsPlatformManager.initialize({
     required MobileTargetPlatform targetPlatform,
@@ -55,17 +72,24 @@ class DefaultsPlatformManager {
   ///
   /// Throws an error if the manager is not initialized.
   static DefaultsPlatformManager get instance {
+    ensureInitialized();
+
+    return _instance!;
+  }
+
+  static bool get isInitialized => _instance != null;
+
+  /// Ensures that the manager has been initialized.
+  ///
+  /// Throws an error if the manager is not initialized.
+  static void ensureInitialized() {
     if (!isInitialized) {
       throw StateError(
         'DefaultsPlatformManager has not been initialized.\n'
         'Call DefaultsPlatformManager.initialize() before accessing instance.',
       );
     }
-
-    return _instance!;
   }
-
-  static bool get isInitialized => _instance != null;
 
   static DefaultsPlatformManager? _instance;
 }